optimize moar

+++ src/js/vec3.ts
@@ -1,7 +1,7 @@
 import { randomMinMax } from './random';
 import { GCVec3Allocator, Vec3Allocator } from './vec3_allocators';
 
-export type Vec3 = Float32Array;
+export type Vec3 = Float64Array;
 export type Color = Vec3;
 export type Point3 = Vec3;
